Just Debugging now

+++ player.java
@@ -1,17 +1,16 @@
 public class player{
     private String name;
-    private int amount;
+    private double amount;
 
     public player(){
         this.name = "User";
         this.amount = 100;
     }
 
-    public player(String name, int amount){
+    public player(String name, double amount){
         this.name = name;
         this.amount = amount;
     }
-
 
     public void setName(String name){
         this.name = name;
@@ -21,11 +20,11 @@ public class player{
         return name;
     }
 
-    public void setAmount(int amount){
+    public void setAmount(double amount){
         this.amount = amount;
     }
 
-    public int getAmount(){
+    public double getAmount(){
         return amount;
     }
 }
